move RichText component to common
