+++ src/common/components/RichText/components/IconButton/index.tsx
