+++ src/common/components/RichText/components/IconButton/styled.tsx
