+++ src/common/components/RichText/utils/isBlockActive/index.tsx
