+++ src/common/components/RichText/utils/isMarkActive/index.ts
